separated profilePage helper methods
